refactor: item helper tool
