+++ app/forms.py
@@ -4,10 +4,6 @@ from wtforms.fields.simple import StringField
 from wtforms.validators import DataRequired, Length
 
 
-class ItemForm(FlaskForm):
-    lost_item = SelectField('Item', choices=[], render_kw={'autofocus': True, 'id': 'lost_item_dropdown'})
-    character = SelectField('Character', choices=[], render_kw={'id': 'character_dropdown'})
-
 class SeedForm(FlaskForm):
     seed1 = SelectField('Seed 1', choices=[], render_kw={'autofocus': True, 'id': 'seed1_dropdown'})
     seed2 = SelectField('Seed 2', choices=[], render_kw={'id': 'seed2_dropdown'})
