+++ tests/test_item_helper.py
@@ -5,7 +5,7 @@ def get_lid(db, name):
     return db.session.query(LostItem).filter_by(name=name).first().id
 
 def assert_response(client, lid, expected_owner, cid, expected_liked_gifts_list):
-    response = client.post('/get_item_data', json={
+    response = client.post('/tools/get_item_data', json={
         'lost_item_selected_option': lid,
         'character_selected_option': cid
     })
